Add files via upload
The app.js file has been updated. The offset which I had for when the player and the enemy collided was .03 at first which was too late. This would give the player more of an advantage. So I have set the offset to .4 which seems like a more logical and makes more sense. 

The README file was updated to include directions on how to play the game locally on your computer.